+++ src/container/index.js
@@ -1,5 +1,5 @@
 export { default as About } from './About/About';
-export { default as Footer } from './Footer/Footer';
+export { default as Contact } from './Contact/Contact';
 export { default as Header } from './Header/Header';
 export { default as Skills } from './Skills/Skills';
 export { default as Testimonials } from './Testimonials/Testimonials';
